# Faster __get__