Benchmarks:

```
+-----------+----------+------------------------------+
| Benchmark | baseline | keyerror                     |
+===========+==========+==============================+
| __get__   | 891 ns   | 646 ns: 1.38x faster (-27%)  |
+-----------+----------+------------------------------+
| __set__   | 28.4 us  | 23.7 us: 1.20x faster (-17%) |
+-----------+----------+------------------------------+
```